+++ upydev/helpinfo.py
@@ -1,0 +1,43 @@
+
+HELP_INFO_ARG ='''Mode/Tools:
+> DEVICE MANAGMENT: '$ upydev dm' to see help on device management.
+    ACTIONS : config, check, set, make_group, mg_group, see, gg
+
+> FILEIO: '$ upydev fio' to see help on file input/ouput operations.
+    ACTIONS: put, get, sync, d_sync, install, update_upyutils
+
+> FIRMWARE: '$ upydev fw' to see help on firmware operations.
+    ACTIONS: fwr, flash, mpyx
+
+> KEYGEN: '$ upydev kg' to see help on keygen operations.
+    ACTIONS: gen_rsakey, rf_wrkey, sslgen_key
+
+> REPLS: '$ upydev rp' to see help on repls modes.
+    ACTIONS: wrepl, wssrepl, srepl
+
+> SHELL-REPLS: '$ upydev sh' to see help on shell-repls modes.
+    ACTIONS: ssl_wrepl, ssl, sh_srepl, shr, wssl, set_wss, ble, jupyterc
+
+> DEBUGGING: '$ upydev db' to see help on debugging operations.
+    ACTIONS: ping, run, timeit, see, find, diagnose, errlog, stream_test,
+             sysctl, log, debug, pytest
+
+> GROUP COMMAND MODE: '$ upydev gp' to see help on group mode options.
+    OPTIONS: -G, -GP
+
+> HELP: '$ upydev h' or '$ upydev help' to see help (without optional args)
+        '$ upydev -h' or '$ upydev --help' to see full help info.
+
+        - see: Help info about ANY ACTION/COMMAND indicated by '-c' option
+
+    ACTIONS: help, h, dm, fio, fw, kg, rp, sh, db, gp, gc, wu, sd, pro.
+
+upy Commands:
+> GENERAL: do '$ upydev gc' to see General commmands help.
+
+> WIFI UTILS: do '$ upydev wu' to see Wifi utils commands help.
+
+> SD: do '$ upydev sd' to see SD utils commands help.
+
+> PROTOTYPE: do '$ upydev pro' to see Prototype utils commands help.
+'''